Add SharedPreferences; store onboarding status
The initial route is now OnboardingStateRoute. When created, it will
display a "loading" page while it checks whether the user has seen the
onboarding page. Depending on the result, it will replace itself with
either HomeRoute or OnboardingRoute.

And of course, OnboardingRoute will save that it has been completed when
it should.